requirements.txt
fastapi
uvicorn
mlflow
scikit-learn
pandas
pydantic
numpy
joblib
databricks-sdk